Написала тесты по условию задачи
Установила пакет testify
