+++ main_test.go
@@ -1,0 +1,65 @@
+package main
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"strings"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestMainHandlerWhenBodyIsNotEmpty(t *testing.T) {
+	req := httptest.NewRequest("GET", "/cafe?count=4&city=moscow", nil)
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	require.Equal(t, http.StatusOK, responseRecorder.Code)
+
+	assert.NotEmpty(t, responseRecorder.Body)
+
+	assert.Equal(t, "Мир кофе,Сладкоежка,Кофе и завтраки,Сытый студент", responseRecorder.Body.String())
+}
+
+func TestMainHandlerWhenCountLessThanTotalCount(t *testing.T) {
+	req := httptest.NewRequest("GET", "/cafe?count=2&city=moscow", nil)
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	require.Equal(t, http.StatusOK, responseRecorder.Code)
+
+	coffeeShops := strings.Split(responseRecorder.Body.String(), ",")
+
+	assert.Len(t, coffeeShops, 2)
+}
+
+func TestMainHandlerWhenCityIsNotSupported(t *testing.T) {
+	req := httptest.NewRequest("GET", "/cafe?count=4&city=kostroma", nil)
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	require.Equal(t, http.StatusBadRequest, responseRecorder.Code)
+
+	assert.Equal(t, "wrong city value", responseRecorder.Body.String())
+}
+
+func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
+	req := httptest.NewRequest("GET", "/cafe?count=7&city=moscow", nil)
+
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
+
+	require.Equal(t, http.StatusOK, responseRecorder.Code)
+
+	coffeeShops := strings.Split(responseRecorder.Body.String(), ",")
+
+	assert.Len(t, coffeeShops, 4)
+}
